% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_constraints.R
\name{adapt_constraints}
\alias{adapt_constraints}
\title{Adapt constraints to a charge input}
\usage{
adapt_constraints(constraints, index)
}
\arguments{
\item{constraints}{a data frame}

\item{index}{a positive integer, the index where a unit of charge is added}
}
\value{
a data frame, updated constraints input
}
\description{
Adapt constraints to a charge input
}
