% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_integer.R
\name{all_integer}
\alias{all_integer}
\title{Check whether all elements of a vector are integer}
\usage{
all_integer(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a logical value
}
\description{
Check whether all elements of a vector are integer
}
