% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_constraints.R
\name{calc_cummax}
\alias{calc_cummax}
\title{Calculates the cumulative maximum charge}
\usage{
calc_cummax(cycles, state, capacity, loss_rate, charge_rate)
}
\arguments{
\item{cycles}{a positive integer}

\item{state}{a positive integer}

\item{capacity}{a positive integer}

\item{loss_rate}{a positive integer}

\item{charge_rate}{a positive integer}
}
\value{
a vector of integers, with length \code{cycles}.
}
\description{
Calculates the cumulative maximum charge
}
