% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_constraints.R
\name{calc_cummin}
\alias{calc_cummin}
\title{Calculates the cumulative minimum charge}
\usage{
calc_cummin(cycles, state, loss_rate)
}
\arguments{
\item{cycles}{a positive integer}

\item{state}{a positive integer}

\item{loss_rate}{a positive integer}
}
\value{
a vector of integers, with length \code{cycles}.
}
\description{
Calculates the cumulative minimum charge
}
