% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_same_length.R
\name{check_same_length}
\alias{check_same_length}
\title{Check all elements for length equality}
\usage{
check_same_length(...)
}
\arguments{
\item{...}{any number of vectors}
}
\value{
TRUE or FALSE
}
\description{
Check all elements for length equality
}
