% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_trades.R
\name{describe_trades}
\alias{describe_trades}
\title{Describe trades}
\usage{
describe_trades(old_schedule, new_schedule, prices)
}
\arguments{
\item{old_schedule}{a numerical vector}

\item{new_schedule}{a numerical vector, must be same length as
\code{current_schedule}}

\item{prices}{a numerical vector}
}
\value{
a data frame with columns for original index, volume and buy/sell price
}
\description{
Generate a trade list base on a previous and a new schedule.
}
\examples{
schedule1 <- c(4, 0, 0,  3)
schedule2 <- c(0, 3, 0, 4)
new_prices <- c(70, 65, 80, 60)
describe_trades(schedule1, schedule2, new_prices)
}
