% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_available_cycles.R
\name{filter_available_cycles}
\alias{filter_available_cycles}
\title{Select only viable charging times}
\usage{
filter_available_cycles(constraints)
}
\arguments{
\item{constraints}{a data frame with columns for cumulative minimum,
\code{cummin}, cumulative maximum \code{cummax}, a direct maximum charge
\code{dirmax}. The order of the rows is assumed to be strictly chronological
without time lapses.

The order of the data frame is assumed to represent the time hierarchy.}
}
\value{
the original data with filtered rows.
}
\description{
A constraints data frame describes the physical limits of a charging process.
}
\details{
The function selects the rows until \code{cummin} column first contains a
value greater than zero (this row is also selected). Out of this selection
only the rows with a \code{cummax} and a \code{dirmax} value greater than
zero are selected.
}
\seealso{
\code{\link{build_constraints}}
}
