% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_da_prices.R
\name{format_da_prices}
\alias{format_da_prices}
\title{Make a day ahead price list}
\usage{
format_da_prices(prices)
}
\arguments{
\item{prices}{a numerical vector of prices.}
}
\value{
a list
}
\description{
Reformats a vector of price data into a day ahead format that is compatible
with \code{optimise_schedule}.
}
\details{
Data is grouped into groups of 24 and put into a list, the first element
containing the first 24 prices, the next 23 elements being empty, the 25th
element containing the next prices and so on.
}
\examples{
some_prices <- rnorm(48, 20, 3)
format_da_prices(some_prices)
}
\seealso{
\code{\link{optimise_schedule}}
}
