% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_id_prices.R
\name{format_id_prices}
\alias{format_id_prices}
\title{Make an intra day price list}
\usage{
format_id_prices(pricetable, colnames)
}
\arguments{
\item{pricetable}{a data frame with columns of price data. The order of the
rows is assumed to be chronological. One row represents one hour.}

\item{colnames}{a vector of strings. describes the names and order of columns
that contain the price data. The first element represents the price one hour
before market closure, the second element describes the price two hours
before market closure and so on. The names cannot be repeated, so each hour
needs to be coded into a separate column.}
}
\value{
the original data frame with the new price list as a further column
}
\description{
Reformats a data frame of index prices into a format compatible with
\code{optimise_schedule}.
}
\details{
Price data is typically formatted so that each row represents a product for a
certain time. For each row there could a column describing the volume
weighted average for the last hour, the last three hours and so on.

With this function, this data is transformed into a list format, where each
element represents a time and contains the current prices at that point in
time for the next hours. The current hour is assumed to be untradeable and
therefore an NA is inserted for that time.
}
\examples{
some_prices <- data.frame(
  time = c(1, 2, 3),
  id_1 = c(20, 30, 40),
  id_2 = c(22, 28, 39),
  id_3 = c(25, 27, 41)
)
format_id_prices(some_prices, c("id_1", "id_2", "id_3"))
}
\seealso{
\code{\link{optimise_schedule}}
}
