% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_non_na_elements.R
\name{identify_non_na_elements}
\alias{identify_non_na_elements}
\title{Find the index of list elements that are not NA}
\usage{
identify_non_na_elements(x)
}
\arguments{
\item{x}{a list}
}
\value{
a vector of the numeric indexes where non-NA values are found
}
\description{
Find the index of list elements that are not NA
}
