% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_constraints.R
\name{match_constraints}
\alias{match_constraints}
\title{Limit constraints to respect reservations}
\usage{
match_constraints(constraints, fixed_schedule, untradeable = NULL)
}
\arguments{
\item{constraints}{A data.frame with constraints}

\item{fixed_schedule}{A numeric vector describing which amounts should be
applied to schedule. It is implicitly assumed that these values do not
exceed the limits of the constraints.}

\item{untradeable}{A numeric vector of indexes in which no trades can occur,
so constraints are adapted accordingly.}
}
\value{
A data.frame of adapted, more limited constraints
}
\description{
This function allows to adapt constraints not only for a single index but
for a range of indexes and a range of values.
}
\examples{
some_constraints <- build_constraints(5, 2, 6, 1, 3)
already_scheduled <- c(0, 2, 2, 1, 0)
untradeable <- c(3, 4)
match_constraints(some_constraints, already_scheduled, untradeable)
}
