% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten.R
\name{shorten}
\alias{shorten}
\title{Shorten a vector}
\usage{
shorten(x, by)
}
\arguments{
\item{x}{a vector.}

\item{by}{an integer value, giving the non-negative number of elements that
should be left out, counted from the end of the vector.}
}
\value{
a vector
}
\description{
Shorten a vector
}
\examples{
vec <- seq(10)
shorten(vec, 3)
}
