% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_marketing.R
\name{simulate_marketing}
\alias{simulate_marketing}
\title{Simulate different marketing scenarios for a set of parameters and prices}
\usage{
simulate_marketing(
  parameters,
  da_prices,
  id_prices,
  id_index_names,
  blocked_per_day = NULL
)
}
\arguments{
\item{parameters}{A named vector of parameters}

\item{da_prices}{A vector of day ahead prices, must be a multiple of 24}

\item{id_prices}{A data frame of intra day prices, different columns describe
different indexes.}

\item{id_index_names}{A vector of strings describing the order of columns 
for consecutive hours.}

\item{blocked_per_day}{An optional vector of 24 integers describing which the
power that is predetermined for certain hours and cannot be optimised.}
}
\value{
A list of results for each of the three scenarios
}
\description{
A single set of parameters, day ahead prices and intra day prices are used
to construct the performance of marketing the parameters without storage on
the day ahead market, with storage on the day ahead market, and with an intra
day marketing on top of day ahead.
}
\seealso{
\code{\link{format_id_prices}}, \code{\link{format_da_prices}},
\code{\link{optimise_schedule}}
}
