% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_no_storage.R
\name{simulate_no_storage}
\alias{simulate_no_storage}
\title{Mock the output \code{optimise_schedule} for settings without storage}
\usage{
simulate_no_storage(parameters, prices)
}
\arguments{
\item{parameters}{a named vector, which must contain \code{loss_rate} and
\code{starting_state}.}

\item{prices}{a numeric vector.}
}
\value{
a list of schedule, state, ans trades. Compatible with
\code{optimise_schedule}
}
\description{
Mock the output \code{optimise_schedule} for settings without storage
}
