% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_schedule.R
\name{split_schedule}
\alias{split_schedule}
\title{Split a schedule in a fixed and a flexible part}
\usage{
split_schedule(schedule, reservations, available_prices)
}
\arguments{
\item{schedule}{A numeric vector of the current schedule}

\item{reservations}{A numeric vector describing reserved amounts of power}

\item{available_prices}{A numeric vector. When an element is NA, this
element is considered to be untradeable, therefore the schedule is taken as
fixed for that index.}
}
\value{
A list of three vectors, the first element describes fixed parts of
the schedule, the second the flexible parts, and a third contains the
indexes where no trade is possible
}
\description{
The parameters schedule, reservations and available_prices should all be
of same length.
}
\examples{
my_schedule <- c(0, 0, 3, 2, 3, 4)
reserved_capacity <- c(TRUE, FALSE, FALSE, FALSE, TRUE, FALSE)
available_prices <- c(30, 40, 20, NA, NA, 5)
split_schedule(my_schedule, reserved_capacity, available_prices)
}
