% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_constraints.R
\name{adapt_constraints}
\alias{adapt_constraints}
\title{Adapt constraints to a charge input}
\usage{
adapt_constraints(constraints, index)
}
\arguments{
\item{constraints}{a data frame of constraints, expected to resemble output
of \code{build_constraints}.}

\item{index}{a positive integer, the index where a unit of charge is added.}
}
\value{
the constraints data frame with updated values
}
\description{
Constraints built with \code{build_constraints} to incorporate the changes
when a charge is added at an index. The values in the columns \code{cummax},
\code{cummin} and \code{dirmax} are updated accordingly.
}
