% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_schedule.R
\name{filter_available_cycles}
\alias{filter_available_cycles}
\title{Select only viable charging times}
\usage{
filter_available_cycles(constraints)
}
\arguments{
\item{constraints}{a data frame with constraints for cumulative minimum,
\code{cummin}, cumulative maximum \code{cummax},
a direct maximum charge \code{dirmax}.

The order of the data frame is assumed to represent the time hierarchy.}
}
\value{
a data frame with original row numbers and prices
}
\description{
Select only viable charging times
}
\seealso{
build_cum_constraints
}
