% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_schedule.R
\name{optimise_schedule}
\alias{optimise_schedule}
\title{Optimise a constrained schedule over price data}
\usage{
optimise_schedule(constraints, prices, volume)
}
\arguments{
\item{constraints}{a data frame with constraints, i.e. columns, for a
cumulative minimum, \code{cummin}, a cumulative maximum \code{cummax}
and a direct maximum charge \code{dirmax}. These}

\item{prices}{a vector of numeric prices, must have same length as rows
of constraints.}

\item{volume}{a positive integer, sets the number of units of energy that
should be distributed over the possible times.}
}
\value{
A vector of integers
}
\description{
The order of the data frame and the price data is assumed to represent
the time hierarchy.
}
\examples{
sample_constraints <- build_constraints(10, 5, 20, 2, 4)
sample_prices <- sample.int(50, 10, replace = TRUE)
optimise_schedule(sample_constraints, sample_prices, 15)
}
