% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_schedule.R
\name{optimise_schedule}
\alias{optimise_schedule}
\title{Optimise a schedule with an iterating approach}
\usage{
optimise_schedule(schedule, prices, parameters, shift, blocked = NULL)
}
\arguments{
\item{schedule}{a numeric vector of the current schedule.}

\item{prices}{a list of available prices. When an element consists only of
NA-values (or a single), then that hour will not be iterated. When some price
values inside a list element are NA, this will be interpreted that only
those hours are not tradeable.}

\item{parameters}{a named vector, including values for charge_rate, loss_rate,
starting_state, and capacity.}

\item{shift}{an integer, indicates the difference to the
the sum of the \code{current_schedule} that is added or subtracted each iteration.}

\item{blocked}{a numeric vector, sane length as \code{schedule},
that blocks amounts that are not optimised but considered when calculating
the needs of the storage.}
}
\value{
A list with three elements, an optimised schedule, the states of the
storage according to that new schedule and a data frame of corresponding trades.
}
\description{
Optimise a schedule with an iterating approach
}
\seealso{
\code{\link{build_constraints}}
}
